--- source/modalSubmissions/_modalSubmission_blueprint.js
@@ -1,9 +1,0 @@
-const ModalSubmission = require('../classes/ModalSubmission.js');
-
-const customId = "";
-module.exports = new ModalSubmission(customId,
-	/** Specs */
-	(interaction, args) => {
-
-	}
-);
